// src/app/generated/cite.api/model/role.ts
/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Evaluation } from './evaluation';
import { Team } from './team';
import { User } from './user';


export interface Role { 
    dateCreated?: Date;
    dateModified?: Date;
    createdBy?: string;
    modifiedBy?: string;
    id?: string;
    evaluationId?: string;
    evaluation?: Evaluation;
    teamId?: string;
    team?: Team;
    name?: string;
    users?: Array<User>;
}